Hackily fix LR(1) conflict in grammar
The grammar is not strictly LR(1) because '{ }' (so two tokens) can
either be an empty record or a pattern in the same context.
The fix reports '{ }' as only one token.

Stranggly enough, '{ a' (which should also conflict) isn't a source of
conflict for menhir. Probably because LALR is somehow more expressive
than LR(1) (assuming menhir managers LALR grammars, which isn't clear).