Translate record patterns to Simple.ast
The default values are translated into optional fields in the pattern to
ease the typing.
If an optional field is absent in the matched record, then the
corresponding variable is given the special value `%%undef`.
Then, let-binding wrapping the body of the function assings the real
default value to the variable.

For example, `{ x ? 1 }: x` will be translated as
```nix
{ x? }:
let %%x = x; in
let x = if x == %%undef then 1 else %%x; in
x
```